src/requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-crispy-forms==1.6.1
facebook-sdk==2.0.0
idna==2.5
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
